refactor: relocate shared scheduler modules to lambda/shared/
Move strategy-agnostic modules that were living in lambda/scheduler/ but
used by both scheduler and reporter (via scheduler_preview):

- sp_types, recommendations, follow_aws_strategy, follow_static_strategy,
  purchase_calculator
- target_strategies/ and split_strategies/ subpackages

They were already being cherry-picked into the reporter zip via explicit
toset(...) in lambda.tf. Relocating to shared/ lets the existing
shared/ fileset handle packaging for all three Lambdas and eliminates
the sys.path hack in reporter/scheduler_preview.py.

Callers updated to use 'from shared.X import ...'. Lambda archive blocks
simplified (removed scheduler-file cherry-pick; added shared sub-package
blocks).
