+++ lambda/scheduler/tests/unit/test_split_strategies.py
@@ -1,8 +1,9 @@
 import pytest
-from split_strategies import calculate_split
-from split_strategies.fixed_step_split import calculate_fixed_step_split
-from split_strategies.gap_split import _resolve_min_purchase, calculate_gap_split
-from split_strategies.one_shot_split import calculate_one_shot_split
+
+from shared.split_strategies import calculate_split
+from shared.split_strategies.fixed_step_split import calculate_fixed_step_split
+from shared.split_strategies.gap_split import _resolve_min_purchase, calculate_gap_split
+from shared.split_strategies.one_shot_split import calculate_one_shot_split
 
 
 # ============================================================================
