+++ lambda/scheduler/tests/unit/test_target_strategies.py
@@ -1,7 +1,8 @@
 import pytest
-from target_strategies import resolve_target
-from target_strategies.aws_target import resolve_aws
-from target_strategies.dynamic_target import resolve_dynamic
+
+from shared.target_strategies import resolve_target
+from shared.target_strategies.aws_target import resolve_aws
+from shared.target_strategies.dynamic_target import resolve_dynamic
 
 
 # ============================================================================
