+++ lambda/shared/follow_static_strategy.py
@@ -3,11 +3,10 @@
 import logging
 from typing import Any
 
-from sp_types import SP_TYPES, get_term
-from split_strategies import calculate_split
-
 from shared.constants import AWS_TYPE_TO_KEY
 from shared.savings_plans_metrics import get_active_savings_plans
+from shared.split_strategies import calculate_split
+from shared.sp_types import SP_TYPES, get_term
 
 
 logger = logging.getLogger()
