+++ lambda/shared/target_strategies/aws_target.py
