+++ lambda/shared/target_strategies/dynamic_target.py
